Extract the cell class into its own module
Move class cell out of grid_tesselate so it sits above the nudge geometry and
per-cell nudge operations in the import order. Confirming a positive-Z contact
builds a throwaway cell and runs zero-height cleanup on it, so the nudge plan
needs the class; leaving it in grid_tesselate would make that a cycle, and
grid_tesselate does not postpone annotation evaluation, so a cycle there is an
import error rather than a soft warning.

Grid tesselation keeps its name and now describes the grid, not the cell.
